requirements.txt
anyio==3.6.2
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cvxpy==1.3.0
ecos==2.0.12
email-validator==1.3.0
exceptiongroup==1.1.0
fastapi==0.88.0
greenlet==2.0.1
h11==0.14.0
Jinja2==3.1.2
matplotlib==3.6.3
numpy==1.24.1
numpy-financial==1.0.0
osqp==0.6.2.post8
packaging==22.0
pandas==1.5.2
passlib==1.7.4
pluggy==1.0.0
pydantic==1.10.4
pyportfolioopt==1.5.4
pytest==7.2.0
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.5
qdldl==0.1.5.post2
requests==2.28.1
scipy==1.10.0
scs==3.2.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
webencodings==0.5.1
yfinance==0.2.3
